Moved Game to Components folder
