+++ code/src/App.js
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux';
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 
 import { games } from './reducers/games';
-import { Game } from './Game';
+import { Game } from './components/Game';
 import { ui } from 'reducers/ui';
 
 const reducer = combineReducers({ games: games.reducer, ui: ui.reducer });
